// output/deploy-scroll-filters-20261010/stable/web-ui/src/types.ts
export interface Bar {
  period_end?: string
  is_closed?: boolean
  datetime: string
  open: number
  high: number
  low: number
  close: number
  vol: number
  amount: number
}